+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Fixed
 - Fix delayed first time invocation of tournament and leaderboard callbacks.
 - Expired tournaments will no longer be listed nor any records will be returned.
+- Unlink device identifiers on console user account details page.
 
 ## [2.5.1] - 2019-05-03
 ### Changed